requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
flake8==6.0.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
snowballstemmer==2.2.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
